change cms

+++ src/count_min_sketch.h
@@ -123,17 +123,17 @@ static __inline__ int cms_import(CountMinSketch *cms, const char *filepath) {
 */
 
 /* Add the provided key to the count-min sketch `x` times */
-int32_t cms_add_inc(CountMinSketch *cms, const char *key, uint32_t x);
-int32_t cms_add_inc_alt(CountMinSketch *cms, uint64_t *hashes,
-                        unsigned int num_hashes, uint32_t x);
+double cms_add_inc(CountMinSketch *cms, const char *key, double x);
+double cms_add_inc_alt(CountMinSketch *cms, uint64_t *hashes,
+                        unsigned int num_hashes, double x);
 
 /* Add the provided key to the count-min sketch */
-static __inline__ int32_t cms_add(CountMinSketch *cms, const char *key) {
+static __inline__ double cms_add(CountMinSketch *cms, const char *key) {
   return cms_add_inc(cms, key, 1);
 }
-static __inline__ int32_t cms_add_alt(CountMinSketch *cms, uint64_t *hashes,
+static __inline__ double cms_add_alt(CountMinSketch *cms, uint64_t *hashes,
                                       unsigned int num_hashes) {
-  return cms_add_inc_alt(cms, hashes, num_hashes, 1);
+  return cms_add_inc_alt(cms, hashes, num_hashes, 1.0);
 }
 
 /*  Remove the provided key to the count-min sketch `x` times;
@@ -155,8 +155,8 @@ static __inline__ int32_t cms_remove_alt(CountMinSketch *cms, uint64_t *hashes,
 }
 
 /* Determine the maximum number of times the key may have been inserted */
-int32_t cms_check(CountMinSketch *cms, const char *key);
-int32_t cms_check_alt(CountMinSketch *cms, uint64_t *hashes,
+double cms_check(CountMinSketch *cms, const char *key);
+double cms_check_alt(CountMinSketch *cms, uint64_t *hashes,
                       unsigned int num_hashes);
 static __inline__ int32_t cms_check_min(CountMinSketch *cms, const char *key) {
   return cms_check(cms, key);
